Moved some functions around.
