+++ src/index.js
@@ -1,8 +1,5 @@
-import { Graph, Edge } from './model';
-import { getSubTree } from './get-sub-tree';
+import { Graph, Edge, getSubTree } from './model';
 import toDot from './to-dot';
-
-export const reverseEdges = ({ edges }) => Array.from(edges).map(({ from, to }) => new Edge(to, from));
 
 export {
   Edge,
